Add filter evaluation for generation evaluation (summarization)
